refactor(xml): move XmlBuilder to its own file and name the scanner's letter range
Review on #432: XmlBuilder moves from scan.ts to libs/xml/src/XmlBuilder.ts,
still internal; the scan() parameter is `builder`, not `slots`, since it
receives the builder itself rather than a copy of its callbacks as the RFC
prototype did; the ASCII letter test that starts an attribute name uses
named character-code constants instead of 64, 91, 96, and 123.

Refs #424
